requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   starlette
    #   watchfiles
bcrypt==4.3.0
    # via passlib
cffi==1.17.1
    # via cryptography
click==8.2.1
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
cryptography==45.0.6
fastapi==0.116.1
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via anyio
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.10
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.33.2
    # via pydantic
pyjwt==2.10.1
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.20
pyyaml==6.0.2
    # via uvicorn
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
starlette==0.47.2
    # via fastapi
typing-extensions==4.14.1
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uvicorn[standard]==0.35.0
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn